fetch and delete existing users

+++ Packages/Account/Sources/Account/AccountViewModel.swift
@@ -56,8 +56,10 @@ import MemosV0Service
         }
 
         // Remove removed users
-        savedUsers.filter { user in !accountManager.accounts.contains { $0.key == user.accountKey } }.forEach { user in
-            currentContext.delete(user)
+        for user in savedUsers {
+            if !accountManager.accounts.contains(where: { $0.key == user.accountKey }) {
+                currentContext.delete(user)
+            }
         }
         try currentContext.save()
         users = allUsers
@@ -83,7 +85,15 @@ import MemosV0Service
         let (user, accessToken) = try await client.signIn(username: username, password: password)
         guard let accessToken = accessToken, let userId = user.id else { throw MoeMemosError.unsupportedVersion }
         let account = Account.memosV1(host: hostURL.absoluteString, id: "\(userId)", accessToken: accessToken)
-        try currentContext.delete(model: User.self, where: #Predicate<User> { user in user.accountKey == account.key })
+        
+        // Fetch and delete existing user if any
+        let predicate = #Predicate<User> { $0.accountKey == account.key }
+        let fetchDescriptor = FetchDescriptor(predicate: predicate)
+        let existingUsers = try currentContext.fetch(fetchDescriptor)
+        for existingUser in existingUsers {
+            currentContext.delete(existingUser)
+        }
+        
         try accountManager.add(account: account)
         try await reloadUsers()
     }
@@ -94,7 +104,15 @@ import MemosV0Service
         let user = try await client.getCurrentUser()
         guard let id = user.remoteId else { throw MoeMemosError.unsupportedVersion }
         let account = Account.memosV1(host: hostURL.absoluteString, id: id, accessToken: accessToken)
-        try currentContext.delete(model: User.self, where: #Predicate<User> { user in user.accountKey == account.key })
+        
+        // Fetch and delete existing user if any
+        let predicate = #Predicate<User> { $0.accountKey == account.key }
+        let fetchDescriptor = FetchDescriptor(predicate: predicate)
+        let existingUsers = try currentContext.fetch(fetchDescriptor)
+        for existingUser in existingUsers {
+            currentContext.delete(existingUser)
+        }
+        
         try accountManager.add(account: account)
         try await reloadUsers()
     }
